Extracted common functions from Phone, Email and Address into Contact
